add pact broker

+++ pact_test_consumer/src/main/java/com/example/pact_test/pact/model/Pizza.java
@@ -12,8 +12,8 @@ import java.util.List;
 public class Pizza {
     private long id;
     private String name;
-    private BigDecimal price;
-    private Integer size;
+    private double price;
+    private int size;
     private List<Topping> toppings;
 
     /**
